requirements.txt
pennylane==0.42.1
autoray==0.6.12
jax==0.4.33
jaxlib==0.4.33
optax==0.2.5
matplotlib==3.10.5
tqdm==4.67.1
aiohttp==3.12.15
h5py==3.14.0
fsspec==2025.7.0
ipykernel==6.30.0

pyyaml==6.0.2
ray==2.49.1
scipy==1.16.1
scikit-image==0.25.2
tensorflow-cpu==2.20.0
tensorflow_datasets==4.9.9
pillow==11.3.0
pandas==2.3.2
scikit-learn==1.7.1
torch==2.8.0
jaxopt==0.8.5
dill==0.4.0
hyperopt==0.2.7
